fix: rename package study to step2
