+++ src/main/kotlin/step2/Calculator.kt
@@ -1,4 +1,4 @@
-package study
+package step2
 
 import java.lang.IllegalArgumentException
