+++ src/test/kotlin/step2/CalculatorTest.kt
@@ -1,4 +1,4 @@
-package study
+package step2
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
